code refactoring for support static sql
